Implement badge listing, remove AllAchievementsHandler
Added badge scraping and display functionality in AllBadgesHandler, including pagination and Steam API integration. Introduced a new Badge model. Removed AllAchievementsHandler and its registration. Updated button handlers and models for consistency with long types for IDs and playtime.

+++ DisSteam/Handlers/AllGamesHandler.cs
@@ -87,7 +87,7 @@ namespace DisSteam.Handlers
 
                 string gameName = gameData.GetProperty("name").ToString();
 
-                int playtimeHours = game.PlayTime / 60;
+                long playtimeHours = game.PlayTime / 60;
 
                 // Get Achievements
                 RestRequest userGameAchievementsRequest = new RestRequest("https://api.steampowered.com/ISteamUserStats/GetPlayerAchievements/v1/", Method.Get)
